Rename sprites-d_isa.png to follow the sprites-<name>3.png convention
Every character spritesheet is sprites-<name>3.png except d_isa, whose file
was sprites-d_isa.png. It worked (imports matched) but violated the
documented convention and was an easy trip-up. Rename the asset and update
its imports.

(kidsfight_scene.backup.ts still references the old name but is dead and
removed separately; TS resolves the wildcard *.png import regardless and it
is not in the build graph.)
